Added a rudiment crypto news feature
Added a new command, `$news`, which fetches the top 5 articles related to cryptocurrencies from the News API organisation. Rather than returning the articles in the same chat, the user receives it in their DM. This makes the chat less cluttered. The articles provided via DMs is cumbersome so it needs a different approach, but this is a start.